chore: embrace clippy and reorganize
