+++ 2023/Rust/Cargo.toml
@@ -12,3 +12,7 @@ readme = "../README.md"
 clap = { version = "4.4.10", features = ["derive"] }
 matches = { version = "0.1.10", features = [] }
 num = "0.4.1"
+
+[[bin]]
+name = "aoc"
+path = "src/main.rs"
