+++ 2023/Rust/src/days/mod.rs
@@ -1,4 +1,5 @@
 pub mod day1;
 pub mod day2;
 
+pub use day1::DayOne;
 pub use day2::DayTwo;
