+++ 2023/Rust/src/lib.rs
@@ -4,18 +4,42 @@ use std::fs;
 use std::path::Path;
 
 pub mod days;
-pub mod problem;
+pub mod puzzle;
 
-pub use problem::Puzzle;
+pub use puzzle::Puzzle;
 
 /// Advent of Code 2023 CLI
 #[derive(Parser)]
 pub struct Cli {
-    /// Day to run (1-25)
-    pub day: usize,
+    /// Day to run (1-25) (Optional, defaults to all available days)
+    pub day: Option<usize>,
     /// Part of the day to run (Optional, defaults to both)
     #[clap(short = 'p', long)]
     pub part: Option<usize>,
+}
+
+pub fn run_day(day: usize, part: Option<usize>) -> Result<(), String> {
+    let input = get_day_input(day).map_err(|e| e.to_string())?;
+
+    let problem: Box<dyn Puzzle> = match day {
+        1 => Box::new(days::day1::DayOne {}),
+        2 => Box::new(days::day2::DayTwo {}),
+        _ => return Err(format!("Day {} not implemented", day)),
+    };
+
+    match part {
+        Some(1) => println!("{}", problem.part_one(&input)),
+        Some(2) => println!("{}", problem.part_two(&input)),
+        None => {
+            println!("{}", problem.name());
+            println!("{}", "-".repeat(problem.name().len()));
+            println!("Part 1: {}", problem.part_one(&input));
+            println!("Part 2: {}", problem.part_two(&input));
+        }
+        _ => return Err("Invalid part".to_string()),
+    };
+
+    Ok(())
 }
 
 /// Get the input from Data/day{day}.input for the given day
