+++ 2023/Rust/src/puzzle.rs
